+++ src/components/wysiwygEditor/decorators/index.js
@@ -1,0 +1,3 @@
+import linkDecorator from './linkDecorator';
+
+export { linkDecorator };